adding context to headers is now part of the agent's private api
