+++ lib/hooks/core/hook-http.js
@@ -63,7 +63,7 @@ function requestWrap(request) {
     if (options.headers) {
       // Adding context to the headers lets us trace the request
       // as it makes it through other layers of the Google infrastructure
-      span.spanData_.span.addContextToHeaders(options.headers);
+      agent.addContextToHeaders(span, options.headers);
     }
 
     var returned = request.call(this, options, function(res) {
