+++ lib/span-data.js
@@ -18,7 +18,6 @@
 
 var TraceSpan = require('./trace-span.js');
 var TraceLabels = require('./trace-labels.js');
-var constants = require('./constants.js');
 
 // Auto-incrementing integer
 var uid = 1;
@@ -80,22 +79,6 @@ SpanData.prototype.close = function() {
   if (this.span.parentSpanId === 0) {
     this.agent.traceWriter.writeSpan(this);
   }
-};
-
-/**
- * Adds the current trace context to the provided http headers
- * so it can follow the associated request through other
- * Google services.
- *
- * @param {Object} headers The http headers associated with the
- *   current request.
- */
-SpanData.prototype.addContextToHeaders = function(headers) {
-  var header = this.trace.traceId + '/' + this.span.spanId;
-  if (this.options) {
-    header += (';o=' + this.options);
-  }
-  headers[constants.TRACE_CONTEXT_HEADER_NAME] = header;
 };
 
 /**
