+++ lib/trace-agent.js
@@ -173,6 +173,27 @@ TraceAgent.prototype.parseContextFromHeader = function(str) {
   };
 };
 
+/**
+ * Adds the provided trace context to the provided http headers
+ * so it can follow the associated request through other
+ * Google services.
+ *
+ * @param {OpaqueSpan} opaque The span to be added to headers.
+ * @param {Object} headers The http headers associated with the
+ *   current request.
+ */
+TraceAgent.prototype.addContextToHeaders = function(opaque, headers) {
+  if (!opaque.spanData_) {
+    return;
+  }
+  var spanData = opaque.spanData_;
+  var header = spanData.trace.traceId + '/' + spanData.span.spanId;
+  if (spanData.options) {
+    header += (';o=' + spanData.options);
+  }
+  headers[constants.TRACE_CONTEXT_HEADER_NAME] = header;
+};
+
 module.exports = {
   get: function(config, logger) {
     if (traceAgent) {
